fix: use default block-action kind so the toggle works and looks native
The custom element rendered a bare uui-button (odd styling, detached from the
action bar) and the click path was unreliable. Follow Umbraco's built-in block
actions (e.g. expose-content): use kind 'default' for the native button and put
all behaviour in the api, exported as 'api'. The api consumes the block entry
context to toggle umbracoNaviHide on execute(), and observes the value to dim the
block (anchored via getHostElement()). Drop the custom element.
